+++ include/Simulation.h
@@ -19,10 +19,10 @@ private:
 public:
     void update(double deltaTime);
     void addWall(int x1, int y1, int x2, int y2);
-    void addParticle(int x, int y, double angle, double velocity);
-    void addParticleByBatchMethod1(int x1, int y1, int x2, int y2, double angle, double velocity, int n);
-    void addParticleByBatchMethod2(int x, int y, double startAngle, double endAngle, double velocity, int n);
-    void addParticleByBatchMethod3(int x, int y, double angle, double startVelocity, double endVelocity, int n);
+    void addParticle(double x, double y, double angle, double velocity);
+    void addParticleByBatchMethod1(double x1, double y1, double x2, double y2, double angle, double velocity, int n);
+    void addParticleByBatchMethod2(double x, double y, double startAngle, double endAngle, double velocity, int n);
+    void addParticleByBatchMethod3(double x, double y, double angle, double startVelocity, double endVelocity, int n);
     std::vector<Particle> getParticles();
 
 };
